pyproject.toml
[tool.poetry]
name = "giobot"
version = "0.1.0"
description = ""
authors = ["devgiordane <bruno_giordane@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^12.8"
httpx = "^0.14.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
